+++ apps/prairielearn/src/tests/activeAccessRestriction.test.sql
@@ -10,17 +10,6 @@ FROM
 WHERE
   u.uid ~ 'student';
 
--- BLOCK select_exam11
-SELECT
-  a.id
-FROM
-  assessments AS a
-  JOIN assessment_sets AS aset ON (aset.id = a.assessment_set_id)
-WHERE
-  a.course_instance_id = 1
-  AND aset.abbreviation = 'E'
-  AND a.number = '11';
-
 -- BLOCK count_variants
 SELECT
   COUNT(v.*)::integer
@@ -30,17 +19,6 @@ FROM
   JOIN assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)
 WHERE
   ai.id = $assessment_instance_id;
-
--- BLOCK select_homework8
-SELECT
-  a.id
-FROM
-  assessments AS a
-  JOIN assessment_sets AS aset ON (aset.id = a.assessment_set_id)
-WHERE
-  a.course_instance_id = 1
-  AND aset.abbreviation = 'HW'
-  AND a.number = '8';
 
 -- BLOCK read_assessment_instance_points
 SELECT